+++ noodles-util/src/alignment/reader.rs
@@ -1,6 +1,6 @@
 //! Alignment reader.
 
-mod builder;
+pub(crate) mod builder;
 
 pub use self::builder::Builder;